Add environ and environ_pass parameters to exec/3
The {environ, EnvMap} option sets values in the environment of the
executed process, using the key of the map as the environment variable
name and the value of the map as the environment variable's value.

The {environ_pass, EnvList} option takes a list of environment variable
names, and takes their values from the current environment to pass into
the subprocess.

If a variable is specified in both options, the environ option takes
precedence.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>